Rename `runtime/compilers/rillv1` to `runtime/parser`
