+++ runtime/parser/data/component-template-v1.json
