+++ runtime/parser/parse_partial_field_selector.go
@@ -1,4 +1,4 @@
-package rillv1
+package parser
 
 import (
 	"errors"
