+++ runtime/parser/parse_report.go
@@ -1,4 +1,4 @@
-package rillv1
+package parser
 
 import (
 	"encoding/json"
